refactor: apply Vercel React best practices to tutorial components
- Use direct imports instead of barrel files (bundle-barrel-imports)
- Store keyboard event handler in refs to avoid re-registering the
  document listener on every prop change (advanced-event-handler-refs)
- Remove no-op content ternary in tooltip
- Use explicit ternary for conditional Back button (rendering-conditional-render)
- Remove unused barrel file
