+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Dashboard } from "@/components/layout/dashboard"
 import { useRemoteControl } from "@/hooks/use-remote-control"
-import { TutorialOverlay } from "@/components/tutorial"
+import { TutorialOverlay } from "@/components/tutorial/tutorial-overlay"
 
 export function App() {
   useRemoteControl()
