--- src/components/tutorial/index.ts
@@ -1,1 +1,0 @@
-export { TutorialOverlay } from "./tutorial-overlay"
