+++ src/components/tutorial/tutorial-tooltip.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useCallback } from "react"
+import { useEffect, useRef } from "react"
 import { SparklesIcon, ChevronLeftIcon } from "lucide-react"
-import type { TooltipRenderProps } from "react-joyride"
+import type { TooltipRenderProps, Controls } from "react-joyride"
 
 export function TutorialTooltip({
   index,
@@ -13,37 +13,41 @@ export function TutorialTooltip({
   controls,
   tooltipProps,
 }: TooltipRenderProps) {
-  const handleKeyDown = useCallback(
-    (e: KeyboardEvent) => {
-      if (e.key === "ArrowRight") {
-        e.preventDefault()
-        if (isLastStep) {
-          controls.skip("button_skip")
-        } else {
-          controls.next()
-        }
-      } else if (e.key === "ArrowLeft" && index > 0) {
-        e.preventDefault()
-        controls.prev()
-      } else if (e.key === "Escape") {
-        e.preventDefault()
-        controls.skip("button_close")
-      }
-    },
-    [controls, index, isLastStep]
-  )
+  const controlsRef = useRef<Controls>(controls)
+  const indexRef = useRef(index)
+  const isLastStepRef = useRef(isLastStep)
+
+  controlsRef.current = controls
+  indexRef.current = index
+  isLastStepRef.current = isLastStep
 
   useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === "ArrowRight") {
+        e.preventDefault()
+        if (isLastStepRef.current) {
+          controlsRef.current.skip("button_skip")
+        } else {
+          controlsRef.current.next()
+        }
+      } else if (e.key === "ArrowLeft" && indexRef.current > 0) {
+        e.preventDefault()
+        controlsRef.current.prev()
+      } else if (e.key === "Escape") {
+        e.preventDefault()
+        controlsRef.current.skip("button_close")
+      }
+    }
+
     document.addEventListener("keydown", handleKeyDown)
     return () => document.removeEventListener("keydown", handleKeyDown)
-  }, [handleKeyDown])
+  }, [])
 
   return (
     <div
       {...tooltipProps}
       className="z-[70] w-[340px] overflow-hidden rounded-xl border border-border/60 bg-card shadow-2xl shadow-black/20"
     >
-      {/* Header with step icon and title */}
       <div className="flex items-center gap-2.5 border-b border-border/40 px-4 pb-3 pt-4">
         <div className="flex size-7 shrink-0 items-center justify-center rounded-lg bg-primary/15">
           <SparklesIcon className="size-3.5 text-primary" />
@@ -53,16 +57,13 @@ export function TutorialTooltip({
         </h3>
       </div>
 
-      {/* Body content */}
       <div className="px-4 py-3">
         <p className="text-[13px] leading-relaxed text-muted-foreground">
           {step.content}
         </p>
       </div>
 
-      {/* Footer: progress + navigation */}
       <div className="flex items-center justify-between border-t border-border/40 px-4 py-3">
-        {/* Progress indicator */}
         <div className="flex items-center gap-1.5">
           {Array.from({ length: size }, (_, i) => (
             <div
@@ -81,7 +82,6 @@ export function TutorialTooltip({
           </span>
         </div>
 
-        {/* Navigation buttons */}
         <div className="flex items-center gap-1.5">
           <button
             {...skipProps}
@@ -89,7 +89,7 @@ export function TutorialTooltip({
           >
             Skip
           </button>
-          {index > 0 && (
+          {index > 0 ? (
             <button
               {...backProps}
               className="inline-flex items-center gap-0.5 rounded-md px-2.5 py-1 text-xs text-muted-foreground transition-colors hover:bg-muted hover:text-foreground"
@@ -97,7 +97,7 @@ export function TutorialTooltip({
               <ChevronLeftIcon className="size-3" />
               Back
             </button>
-          )}
+          ) : null}
           <button
             {...primaryProps}
             className="rounded-md bg-primary px-3 py-1 text-xs font-medium text-primary-foreground transition-colors hover:bg-primary/90"
